+++ magpie/adapter/magpieservice.py
@@ -1,6 +1,7 @@
 """
 Store adapters to read data from magpie.
 """
+from distutils.version import LooseVersion
 from typing import TYPE_CHECKING
 
 import requests
@@ -11,15 +12,34 @@ from pyramid.settings import asbool
 from magpie.api.schemas import ServicesAPI
 from magpie.models import Service as MagpieService
 from magpie.services import invalidate_service
-from magpie.utils import CONTENT_TYPE_JSON, get_admin_cookies, get_logger, get_magpie_url, get_settings
+from magpie.utils import (
+    CONTENT_TYPE_JSON,
+    get_admin_cookies,
+    get_logger,
+    get_magpie_url,
+    get_settings,
+    get_twitcher_url
+)
 
 # WARNING:
 #   Twitcher available only when this module is imported from it.
 #   It is installed during tests for evaluation.
 #   Module 'magpie.adapter' should not be imported from 'magpie' package.
-from twitcher.datatype import Service as TwitcherService  # noqa
+from twitcher.__version__ import __version__ as twitcher_version  # noqa
 from twitcher.exceptions import ServiceNotFound  # noqa
-from twitcher.store import ServiceStoreInterface  # noqa
+
+if LooseVersion(twitcher_version) > LooseVersion("0.6.0"):
+    from twitcher.models import Service as TwitcherService  # noqa  # pylint: disable=E0611  # Twitcher >= 0.6.0
+    from twitcher.store import ServiceStoreInterface  # noqa  # pylint: disable=E0611  # Twitcher > 0.6.0
+elif LooseVersion(twitcher_version) == LooseVersion("0.6.0"):
+    from twitcher.models import Service as TwitcherService  # noqa  # pylint: disable=E0611  # Twitcher >= 0.6.0
+
+    class ServiceStoreInterface(object):  # was removed on initial 0.6.0 version
+        def __init__(self, request):
+            self.request = request
+else:
+    from twitcher.datatype import Service as TwitcherService  # noqa  # pylint: disable=E0611  # Twitcher <= 0.5.x
+    from twitcher.store import ServiceStoreInterface  # noqa  # pylint: disable=E0611  # Twitcher <= 0.5.x
 
 if TYPE_CHECKING:
     from pyramid.request import Request
@@ -43,24 +63,42 @@ class MagpieServiceStore(ServiceStoreInterface):
         self.settings = get_settings(request)
         self.session_factory = request.registry["dbsession_factory"]
         self.magpie_url = get_magpie_url(request)
+        self.twitcher_url = get_twitcher_url(request)
         self.twitcher_ssl_verify = asbool(self.settings.get("twitcher.ows_proxy_ssl_verify", True))
         self.magpie_admin_token = get_admin_cookies(self.settings, self.twitcher_ssl_verify)
 
-    def save_service(self, service, overwrite=True, request=None):
+    def save_service(self, service=None, name=None, url=None, overwrite=True, request=None):
         """
-        Magpie store is read-only, use magpie api to add services.
+        Store is read-only, use `Magpie` :term:`API` to add services.
+
+        .. note::
+            Multiple redundant parameters are applied to support different `Twitcher` versions.
+
+            - ``Twitcher <=0.5.x`` uses ``(service, *args, **kwargs)``
+            - ``Twitcher >=0.6.x`` uses ``(name, url, *args, **kwargs)``
+            - Some alternate interfaces also provided extra parameters at some point.
         """
-        raise NotImplementedError
+        msg = (
+            "MagpieAdapter does not support 'MagpieServiceStore.save_service' operation. "
+            "Use Magpie API or UI for service registration."
+        )
+        LOGGER.error(msg)
+        raise NotImplementedError(msg)
 
     def delete_service(self, name, request=None):
         """
-        Magpie store is read-only, use magpie api to delete services.
+        Store is read-only, use :mod:`Magpie` :term:`API` to delete services.
         """
-        raise NotImplementedError
+        msg = (
+            "MagpieAdapter does not support 'MagpieServiceStore.delete_service' operation. "
+            "Use Magpie API or UI for service removal."
+        )
+        LOGGER.error(msg)
+        raise NotImplementedError(msg)
 
     def list_services(self, request=None):  # noqa: F811
         """
-        Lists all services registered in magpie.
+        Lists all services registered in `Magpie`.
         """
         # obtain admin access since 'service_url' is only provided on admin routes
         services = []
@@ -100,19 +138,22 @@ class MagpieServiceStore(ServiceStoreInterface):
             - :meth:`magpie.adapter.magpieowssecurity.MagpieOWSSecurity.get_service`
             - :meth:`magpie.adapter.magpieservice.MagpieServiceStore.fetch_by_name`
         """
-        session = self.session_factory()
-
-        try:
-            service = MagpieService.by_service_name(service_name, db_session=session)
-            if service is None:
-                raise ServiceNotFound("Service name not found.")
-
-            return TwitcherService(url=service.url,
-                                   name=service.resource_name,
-                                   type=service.type,
-                                   verify=self.twitcher_ssl_verify)
-        finally:
-            session.close()
+        session = self.request.db
+        service = MagpieService.by_service_name(service_name, db_session=session)
+        if service is None:
+            raise ServiceNotFound("Service name not found.")
+        service_data = dict(
+            url=service.url,
+            name=service.resource_name,
+            type=service.type,
+            verify=self.twitcher_ssl_verify
+        )
+        if LooseVersion(twitcher_version) >= LooseVersion("0.6.0"):
+            service_data["_verify"] = int(service_data.pop("verify"))  # parameter renamed and different type
+            service_data["id"] = service.resource_id
+            service_data["purl"] = "{}/{}".format(self.twitcher_url, service.resource_name)
+            service_data["auth"] = "magpie"
+        return TwitcherService(**service_data)
 
     def fetch_by_name(self, name):
         # type: (Str) -> TwitcherService
@@ -124,7 +165,8 @@ class MagpieServiceStore(ServiceStoreInterface):
             cache_regions["service"] = {"enabled": False}
         if self.request.headers.get("Cache-Control") == "no-cache":
             invalidate_service(name)
-        return self._fetch_by_name_cached(name)
+        service = self._fetch_by_name_cached(name)
+        return service
 
     def fetch_by_url(self, url, request=None):
         """